// database/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export interface Database {
  public: {
    Tables: {
// ...
      seasons: {
        Row: {
          id: string;
          name: string;
          start_at: string;
          end_at: string;
          is_closed: boolean;
          created_at: string;
          updated_at: string;
        };
        Insert: {
          id?: string;
          name: string;
          start_at: string;
          end_at: string;
          is_closed?: boolean;
          created_at?: string;
          updated_at?: string;
        };
        Update: {
          id?: string;
          name?: string;
          start_at?: string;
          end_at?: string;
          is_closed?: boolean;
          created_at?: string;
          updated_at?: string;
        };
        Relationships: [];
      };
// ...
      season_snapshots: {
        Row: {
          id: string;
          season_id: string;
          snapshot: any;
          created_at: string;
        };
        Insert: {
          id?: string;
          season_id: string;
          snapshot: any;
          created_at?: string;
        };
        Update: {
          id?: string;
          season_id?: string;
          snapshot?: Json;
          created_at?: string;
        };
        Relationships: [];
      };
    };
// ...
  };
}